Experience Sampling Django/experiencesampling/requirements.txt
Django==2.2.7
channels==2.3.1
channels_redis==2.4.1
Pillow==6.2.1
pytest-django==3.6.0
pytest-asyncio==0.10.0
djangorestframework==3.10.3
django-apscheduler
apscheduler